==> 01_volatility_recover.Rmd <==
---
title: "Volatility and Informativeness"
author: "Eduardo Davila^[Yale] & Cecilia Parlatore^[NYU Stern]"
date: "`r Sys.Date()`"
output:
  html_document: default
  pdf_document: default
---

```{r echo=FALSE, include=FALSE}

library(here); library(tidyverse); library(foreach); library(doParallel); library(AER)
path <- here::here(); print(path); setwd(path); rm(path)

source("functions/fn_identifying_stocks.R")
source("functions/fn_recover.R")
source("functions/fn_rolling_stocks.R")

load("input/data_selected.RData")

cl <- makeCluster(11); registerDoParallel(cl)

```

# Recovering Kalman Gains 

Quarterly observations

```{r}

payoffname <- "growth"

controls_levels <- c()
controls_logs   <- c()

df_input   <- df_q
month_fe   <- TRUE
window_T_q <- 40 # must be less or equal than N_q or N_a

results_rolling_q <- fn_rolling_stocks_par(df_input, payoffname, controls_levels, controls_logs, window_T_q, month_fe)
# results_rolling_q <- fn_rolling_stocks(df_input, payoffname, controls_levels, controls_logs, window_T_q, month_fe)

save(results_rolling_q, file = "intermediate/results_rolling.RData")

```
