==> 04_volatility_figure.Rmd <==
---
title: "Volatility and Informativeness"
author: "Eduardo Davila^[Yale] & Cecilia Parlatore^[NYU Stern]"
date: "`r Sys.Date()`"
output:
  html_document: default
  pdf_document: default

```{r echo=FALSE, include=FALSE}

library(here); library(tidyverse);library(kableExtra); library(latex2exp); options(dplyr.summarise.inform = F)
path <- here::here(); print(path); setwd(path); rm(path)

load("intermediate/results_rolling_clean.RData")

```

# FIGURE

```{r}

limit <- 5

fun.1 <- function(x) {(sqrt(1 + 8*x) - 1 + 2*x)/(2*x^2)}
fun.2 <- function(x) {(x - 2 + sqrt((2 - x)^2 + 8*x^2))/(4*x^2)}

g1 <- ggplot(data = data.frame(x = c(0,limit)),
             mapping = aes(x = x)) +
  stat_function(fun = fun.1, color = "darkgoldenrod2", size = 1.8, n = 1000) + 
  stat_function(fun = fun.2, color = "darkgoldenrod2", size = 1.8, n = 1000) +
  theme(axis.ticks.length = unit(-0.15, "cm"),
        axis.text.x = element_text(color = "black", family = "serif", size = 10, margin = margin(t = .5, unit = "cm")),
        axis.text.y = element_text(color = "black", family = "serif", size = 10, margin = margin(r = .5, unit = "cm")),    
        axis.title  = element_text(family = "serif", size = 14),
        panel.background = element_blank(),
        panel.grid.major = element_line(colour = "darkgray", size = 0.05),
        panel.grid.minor = element_line(colour = "darkgray", size = 0.05),
        axis.line = element_line(colour = "black")) +
  scale_x_continuous(minor_breaks = seq(1, limit, 1), expand = c(0,0), limits = c(0,limit), sec.axis = dup_axis(name = "", labels = NULL)) +
  scale_y_continuous(minor_breaks = seq(1, limit, 1), expand = c(0,0), limits = c(0,limit), sec.axis = dup_axis(name = "", labels = NULL)) +
  xlab(TeX("$\\frac{\\tau_s}{\\tau_{\\eta}}$ (Signal-to-payoff)")) +
  ylab(TeX("$\\frac{\\tau_n}{\\tau_{\\eta}}$ (Noise-to-payoff)")) + 
  annotate("text", family = "serif", color = "black", fontface = "plain", x = 2.2,  y = 1.8, label = "Positive Comovement")  + 
  annotate("text", family = "serif", color = "black", fontface = "plain", x = 1.05, y = .2, label = "Negative Comovement") + 
  annotate("text", family = "serif", color = "black", fontface = "plain", x = 1.01, y = .8,  label = "Ambiguous Comovement")

xs <- seq(0, 15, length.out = 1000)
ysmin <- fun.1(xs)
ysmax <- fun.2(xs)
df2 <- data.frame(xs, ysmin, ysmax)

df_year <- results %>% filter(year_end == 2000)

directory <- "output"

ggsave(paste(directory, "/", "regions", ".pdf", sep = ""), plot = g1, width = 5, height = 5, units = "in", dpi = 300)

```
